+++ spec/apps/rails/dummy_app.rb
@@ -32,6 +32,7 @@ class DummyApp < Rails::Application
     get '/crash' => 'dummy#crash'
     get '/breakdown' => 'dummy#breakdown'
     get '/breakdown_view_only' => 'dummy#breakdown_view_only'
+    get '/breakdown_http' => 'dummy#breakdown_http'
     get '/notify_airbrake_helper' => 'dummy#notify_airbrake_helper'
     get '/notify_airbrake_sync_helper' => 'dummy#notify_airbrake_sync_helper'
     get '/active_record_after_commit' => 'dummy#active_record_after_commit'
@@ -104,7 +105,8 @@ class DummyController < ActionController::Base
       'dummy/resque.html.erb' => 'resque',
       'dummy/delayed_job.html.erb' => 'delayed_job',
       'dummy/breakdown.html.erb' => 'breakdown',
-      'dummy/breakdown_view_only.html.erb' => 'breakdown_view_only'
+      'dummy/breakdown_view_only.html.erb' => 'breakdown_view_only',
+      'dummy/breakdown_http.html.erb' => 'breakdown_http'
     )
   ]
 
@@ -122,6 +124,11 @@ class DummyController < ActionController::Base
 
   def breakdown_view_only
     render 'dummy/breakdown.html.erb'
+  end
+
+  def breakdown_http
+    Net::HTTP.get('example.com', '/')
+    render 'dummy/breakdown_http.html.erb'
   end
 
   def notify_airbrake_helper