+++ spec/integration/rails/rails_spec.rb
@@ -321,5 +321,18 @@ RSpec.describe "Rails integration specs" do
         get '/breakdown_view_only'
       end
     end
+
+    context "when an action performs a HTTP request" do
+      let!(:example_request) do
+        stub_request(:get, 'http://example.com').to_return(body: '')
+      end
+
+      it "includes the http breakdown" do
+        expect(Airbrake).to receive(:notify_performance_breakdown)
+          .with(hash_including(groups: { view: be > 0, http: be > 0 }))
+        get '/breakdown_http'
+        expect(example_request).to have_been_made
+      end
+    end
   end
 end